Encoder: Fix in returning recon buffers with smaller IDR interval
For every IDR generated after the first one, one recon buffer was not released
from the buffer manager, when recon was enabled.
This resulted in encoder returning with an error after couple of IDRs when
recon was enabled.
This is fixed by calling recon buffer release based on pic_cnt instead of frm_num.
frm_num is reset to 0 for every IDR where as pic_cnt is not.

Change-Id: I8b72e5aedc6a460292388e4e46f56c07486db4a6
Signed-off-by: mydongistiny <jaysonedson@gmail.com>